% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{cal_curve_MSI}
\alias{cal_curve_MSI}
\title{Dataset cal_curve_MSI}
\description{
This is an DESI-MRM dataset (in imzML format) containing tissue and calibration data as well as the linera models for quantification.
}
\section{tissue_MRM_data.raw/cal_curve_MSI.RDS}{


This data is used to test int2conc() function.
}

