% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{cal_MRM_data}
\alias{cal_MRM_data}
\title{Dataset cal_MRM_data}
\description{
This is a dataset containing Waters MRM data of SphingoMyelin 16:00 in Guinea pig lung tissue.
}
\section{cal_MRM_data.raw/imaging/Analyte 1.txt}{


This data is used in read_mrm().
}

