% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{cal_response_data}
\alias{cal_response_data}
\title{Dataset cal_response_data}
\description{
This is a table containing mean intensity per pixel for the ROI at each calibration level across all calibration replicates.
}
\section{cal_response_data.csv}{


This data is to test the create_cal_curve() function.
}

