% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{cal_rois}
\alias{cal_rois}
\title{Dataset cal_rois}
\description{
This is a table containing a logical data about caliobration levels pixels are associated with.
}
\section{cal_MRM_data.raw/cal_ROIs.csv}{


This data is used to identify calibration spots.
}

