% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{calibration_metadata}
\alias{calibration_metadata}
\title{Dataset calibration_metadata}
\description{
This is a table containing information about concentrations in the calibration spots.
}
\section{calibration_metadata.csv}{


This data is used to create calibration curves in create_cal_curve().
}

