% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_MSIs.R
\name{combine_MSIs,MSImagingExperiment-method}
\alias{combine_MSIs,MSImagingExperiment-method}
\title{Function to combine MSImagingExperiment objects.}
\usage{
\S4method{combine_MSIs}{MSImagingExperiment}(MSIobject, ...)
}
\arguments{
\item{MSIobject}{MSImagingExperiment object from Cardinal}

\item{...}{additional MSImagingExperiment object to combine - must have matching fData() and same columns form pData().}
}
\value{
quant_MSImagingExperiment object with intensity values replaced with response
}
\description{
Function to combine MSImagingExperiment objects.
}
