% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{combined}
\alias{combined}
\title{Dataset combined}
\description{
This is an DESI-MRM dataset (in imzML format) containing tissue and calibration data.
}
\section{tissue_MRM_data.raw/combined.imzML}{


This data is used to test the createMSIDatamatrix() summarise_cal_levels() and create_cal_curve() functions.
}

