% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{createClassAndColors}
\alias{createClassAndColors}
\title{Ggplot2 pal for publication figures}
\usage{
createClassAndColors(
  class,
  QC_label = "QC",
  Blank_label = "Blank",
  QC_color = "#000000",
  Blank_color = "#A65628",
  manual_color = c("#386cb0", "#ef3b2c", "#7fc97f", "#fdb462", "#984ea3", "#a6cee3",
    "#778899", "#fb9a99", "#ffff33")
)
}
\description{
Ggplot2 pal for publication figures
}
