% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMSIDatamatrix.R
\name{createMSIDatamatrix,quant_MSImagingExperiment-method}
\alias{createMSIDatamatrix,quant_MSImagingExperiment-method}
\title{Function to create data matrix from MSI object}
\usage{
\S4method{createMSIDatamatrix}{quant_MSImagingExperiment}(MSIobject, inputNA = T)
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}

\item{inputNA}{Whether to convert 0's in matrix to NA (default = T)}
}
\value{
MSIobject with slots updated for i) matrix of average ng/pixel of m/z (rows = m/z and cols = cal level) in tissue ROIs ii) sample/ROI metadata
}
\description{
Function to create data matrix from MSI object
}
