% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{example_MRM_raw}
\alias{example_MRM_raw}
\title{Dataset example_MRM_raw}
\description{
This is a dataset containing Waters MRM data of SphingoMyelin 16:00 in Guinea pig lung tissue
}
\section{example_MRM_raw.raw/imaging/Analyte 1.txt}{


This data is used in read_mrm().
}

