% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2conc.R
\name{int2conc,quant_MSImagingExperiment-method}
\alias{int2conc,quant_MSImagingExperiment-method}
\title{Function to update intensity with concentration values}
\usage{
\S4method{int2conc}{quant_MSImagingExperiment}(MSIobject, cal_label = "Cal")
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}

\item{cal_label}{Label in pixel metadata which corresponds to calibration data}
}
\value{
MSIobject with intensity values replaced with concentration values (ng/pixel)
}
\description{
Function to update intensity with concentration values
}
