% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2response.R
\name{int2response,quant_MSImagingExperiment-method}
\alias{int2response,quant_MSImagingExperiment-method}
\title{Function to normalise the intensity values to response per pixel if internal standard is present.}
\usage{
\S4method{int2response}{quant_MSImagingExperiment}(MSIobject)
}
\arguments{
\item{MSIobject}{MSI MSIobject from Cardinal}
}
\value{
MSIobject with intensity values replaced with response
}
\description{
Function to normalise the intensity values to response per pixel if internal standard is present.
}
