% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{ion_library}
\alias{ion_library}
\title{Dataset ion_library}
\description{
This is a table containing a library of MRM transitions included in targeted MSI experiments.
}
\section{ion_library.txt}{


This data is used in read_mrm().
}

