% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mrm.R
\name{read_mrm}
\alias{read_mrm}
\title{Function to create data matrix from MSI object}
\usage{
read_mrm(name, folder)
}
\arguments{
\item{name}{Name of Waters raw imaging file (including the imaging/Analyte 1.txt file after processing)}

\item{folder}{Location of folder with raw imaging data in}
}
\value{
MSIobject with slots updated for i) matrix of average ng/pixel of m/z (rows = m/z and cols = cal level) in tissue ROIs ii) sample/ROI metadata
}
\description{
Function to create data matrix from MSI object
}
