% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{sample_metadata}
\alias{sample_metadata}
\title{Dataset calibration_metadata}
\description{
This is a table containing information about the study sample(s).
}
\section{sample_metadata.csv}{


This data is used to update the pData() of MSIobjects.
}

