% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setCommonAxis.R
\name{setCommonAxis,list-method}
\alias{setCommonAxis,list-method}
\title{Function to ser list of MSImagingExperiment objects set to same feature axis}
\usage{
\S4method{setCommonAxis}{list}(MSIobjects, ref_fdata)
}
\arguments{
\item{MSIobjects}{List of MSImagingExperiment objects}

\item{ref_fdata}{MassDataFrame with features to use}
}
\value{
List of MSIobjects with common fData and matching spectra
}
\description{
Function to ser list of MSImagingExperiment objects set to same feature axis
}
