% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{Slide12}
\alias{Slide12}
\title{Microscope image of H&E stain (.czi)}
\description{
Microscope image of H&E stain (.czi) to extract image coordinates
}
\section{Slide12.czi}{


This data is used to test colocalising images
}

