% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{Slide12_shrunk}
\alias{Slide12_shrunk}
\title{Shrunk microscope image of H&E stain (.png)}
\description{
Shrunk image of H&E stain (.png) to colocalise with ion image
}
\section{Slide12_shrunk.png}{


This data is used to test colocalising images
}

