% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{snr_image}
\alias{snr_image}
\title{Ion image (.png)}
\description{
Ion image of 12,13_DiHOME (.png) to colocalise with H&E stain
}
\section{snr_image.png}{


This data is used to test colocalising images
}

