% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_cal_levels.R
\name{summarise_cal_levels,quant_MSImagingExperiment-method}
\alias{summarise_cal_levels,quant_MSImagingExperiment-method}
\title{Function to calculate the mean response or intensity per pixel for the ROI at each calibration level across all calibration replicates (ng/pixel).}
\usage{
\S4method{summarise_cal_levels}{quant_MSImagingExperiment}(MSIobject, cal_label = "Cal")
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}

\item{cal_label}{Label in pixel metadata which corresponds to calibration data}
}
\value{
MSIobject with slots updated for i) matrix of average ng/pixel of m/z (rows = m/z and cols = cal level) ii) list of pixel counts per cal level
}
\description{
Function to calculate the mean response or intensity per pixel for the ROI at each calibration level across all calibration replicates (ng/pixel).
}
