% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_themes.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Ggplot2 theme for publication figures}
\usage{
theme_Publication(base_size = 14)
}
\arguments{
\item{base_size}{Set base size}
}
\description{
Ggplot2 theme for publication figures
}
