% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{tissue_pixels}
\alias{tissue_pixels}
\title{Dataset tissue_pixels}
\description{
This is a table containing a logical data about whether pixel is part of a calibration level.
}
\section{tissue_pixels.csv}{


This data is used to identify tissue pixels.
}

