% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extdata.R
\name{tissue_rois}
\alias{tissue_rois}
\title{Dataset tissue_rois}
\description{
This is a table containing a logical data about which tissue types pixels are associated with.
}
\section{tissue_rois.csv}{


This data is used to identify and label tissue types.
}

