% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero2na.R
\name{zero2na,quant_MSImagingExperiment-method}
\alias{zero2na,quant_MSImagingExperiment-method}
\title{Function to convert intensity values from 0 to NA in MSI dataset.}
\usage{
\S4method{zero2na}{quant_MSImagingExperiment}(MSIobject, val_slot = "intensity")
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}
}
\value{
MSIobject with intensity values replaced with response
}
\description{
Function to convert intensity values from 0 to NA in MSI dataset.
}
