% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_MSIs.R
\name{combine_MSIs,MSImagingExperiment-method}
\alias{combine_MSIs,MSImagingExperiment-method}
\title{Function to convert intensity values from 0 to NA in MSI dataset.}
\usage{
\S4method{combine_MSIs}{MSImagingExperiment}(MSIobject, ...)
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}

\item{...}{additional MSI objects to combine}
}
\value{
MSIobject with intensity values replaced with response
}
\description{
Function to convert intensity values from 0 to NA in MSI dataset.
}
