% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createMSIDatamatrix.R
\name{createMSIDatamatrix,quant_MSImagingExperiment-method}
\alias{createMSIDatamatrix,quant_MSImagingExperiment-method}
\title{Function to create data matrix from MSI object}
\usage{
\S4method{createMSIDatamatrix}{quant_MSImagingExperiment}(
  MSIobject,
  val_slot = "intensity",
  inputNA = T,
  roi_header = NA
)
}
\arguments{
\item{MSIobject}{MSI object from Cardinal, pData() to include sample_ID.}

\item{val_slot}{character defining slot name to normalise - takes "intensity" as default}

\item{inputNA}{Whether to convert 0's in matrix to NA (default = T)}

\item{roi_header}{Header in pData pertaining to ROIs to average. Set to NA to skip generating average df}
}
\value{
MSIobject with slots updated for i) matrix of average ng/pixel of m/z (rows = m/z and cols = cal level) in tissue ROIs ii) sample/ROI metadata
}
\description{
Function to create data matrix from MSI object
}
