% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2conc.R
\name{int2conc,quant_MSImagingExperiment-method}
\alias{int2conc,quant_MSImagingExperiment-method}
\title{Function to update intensity with concentration values}
\usage{
\S4method{int2conc}{quant_MSImagingExperiment}(MSIobject, val_slot = "response", pixels = "Tissue")
}
\arguments{
\item{MSIobject}{MSI object from Cardinal}

\item{val_slot}{character defining slot name to normalise - takes "intensity" as default}

\item{pixels}{Label in pixel metadata the pixels to quantify}
}
\value{
MSIobject with intensity values replaced with concentration values (ng/pixel)
}
\description{
Function to update intensity with concentration values
}
