% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2response.R
\name{int2response,quant_MSImagingExperiment-method}
\alias{int2response,quant_MSImagingExperiment-method}
\title{Function to normalise the intensity values to response per pixel if internal standard is present. Currently only works for a single internal standard to normalise all lipids to.}
\usage{
\S4method{int2response}{quant_MSImagingExperiment}(
  MSIobject,
  val_slot = "intensity",
  IS_name = "None",
  mode = "line",
  remove_IS = T,
  ...
)
}
\arguments{
\item{MSIobject}{MSI MSIobject from Cardinal}

\item{val_slot}{character defining slot name to normalise - takes "intensity" as default}

\item{IS_name}{character defining the IS to use in fData(MSIobject) under the analyte header. Currently only works for a single internal standard to normalise all lipids to. "None" use individual lipids to normalise.}

\item{mode}{Mode iby which to apply normalisation. "line" = normalise to median intensity of IS per sample, "line" = normalise to median intensity of IS per line, "pixel" = normalise to intensity of IS per pixel.}

\item{remove_IS}{Logical whether to remove internal standard feature from object}
}
\value{
MSIobject with intensity values replaced with response
}
\description{
Function to normalise the intensity values to response per pixel if internal standard is present. Currently only works for a single internal standard to normalise all lipids to.
}
