% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int2snr.R
\name{int2snr,quant_MSImagingExperiment-method}
\alias{int2snr,quant_MSImagingExperiment-method}
\title{Function to convert the intensity values to SNR per pixel based on same transitions in noise/background pixels.Run after IS normalization.}
\usage{
\S4method{int2snr}{quant_MSImagingExperiment}(MSIobject, noise = "Noise", tissue = "Tissue", snr_thresh = 3, ..)
}
\arguments{
\item{MSIobject}{MSI MSIobject from Cardinal - including pData(MSIobject)$sample_type == Noise}

\item{noise}{character in pData(MSIobject)$sample_type which indicates background / noise pixels}

\item{tissue}{character in pData(MSIobject)$sample_type which indicates tissue pixels to calculate SNR for}

\item{snr_thresh}{value indicating the minimum SNR value to accept (below this value SNR = 0)}
}
\value{
MSIobject with intensity values replaced with SNR values
}
\description{
Function to convert the intensity values to SNR per pixel based on same transitions in noise/background pixels.Run after IS normalization.
}
